Remove old Storage files
