+++ Source/Shared/Library/Entry.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 /// A wrapper around cached object and its expiry date.
-public struct Entry<T: Codable> {
+public struct Entry<T> {
   /// Cached object
   public let object: T
   /// Expiry date
